worker/requirements.txt
numpy==2.2.4
onnxruntime==1.21.0
opencv-python-headless==4.11.0.86
Pillow==11.1.0
scikit-learn==1.6.1
supabase==2.15.0
insightface==0.7.3

